Extract custom widgets and helpers
