+++ lapce-ui/src/lib.rs
@@ -27,3 +27,5 @@ mod tab;
 pub mod terminal;
 pub mod title;
 pub mod window;
+
+pub mod widgets;
